+++ Lab4/src/Logger.cpp
@@ -25,11 +25,11 @@ void Logger::draw(sf::RenderWindow& window) {
 }
 
 void Logger::update() {
-    double fps = m_counter->get_fps();
+    float fps = m_counter->get_fps();
     Vector4 eye = m_camera->get_eye();
     Vector4 target = m_camera->get_target();
     Vector4 up = m_camera->get_up();
-    double scale = m_camera->get_scale();
+    float scale = m_camera->get_scale();
 
     std::string text_str = std::format(
         "FPS = {:.0f}\n"
